Background fixes for notification service
